requirements.txt
pyrogram
tgcrypto
python-dotenv
pyTelegramBotAPI
asyncio
